requirements/local.txt
-r base.txt

# A configurable set of panels that display various debug information about
# the current request/response.
# http://django-debug-toolbar.readthedocs.org/
django-debug-toolbar >= 1.4, < 1.5

# The Python WSGI Utility Library
# http://werkzeug.pocoo.org/
Werkzeug >= 0.11, < 0.12